fix timedelta problem by using pytz and add redirect if remaining_amount is not enough in reservsation process

+++ app/schemas/customer.py
@@ -20,7 +20,7 @@ class CustomerCreate(CustomerBase):
 class CustomerUpdate(BaseModel):
     subscription_model: Optional[str] = Field(None, description="Subscription model of the customer", example="premium")
     subscription_end_time: Optional[datetime] = Field(None, description="End time of the subscription", example="2024-12-31T23:59:59")
-    wallet_money_amount: Optional[float] = Field(None, description="Amount of money in the wallet", example=20000)
+    wallet_money_amount: Optional[int] = Field(None, description="Amount of money in the wallet", example=20000)
 
 class CustomerOut(CustomerBase):
     id: int = Field(..., description="The unique identifier for the customer", example=1)
